Move file its in
