+++ x/gamm/pool-models/balancer/amm_test.go
@@ -2,9 +2,7 @@ package balancer_test
 
 import (
 	"fmt"
-	"math/rand"
 	"testing"
-	time "time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
@@ -342,115 +340,5 @@ func TestCalcSingleAssetInAndOut_InverseRelationship(t *testing.T) {
 				require.True(osmoutils.DecApproxEq(t, initialCalcTokenOut.ToDec(), inverseCalcTokenOut, tol))
 			})
 		}
-	}
-}
-
-// Tests selecting a random amount of coins to LP, and then that ExitPool(JoinPool(tokens))
-// preserves the pools number of LP shares, and returns fewer coins to the acter than they started with.
-func TestRandomizedJoinPoolExitPoolInvariants(t *testing.T) {
-	type testCase struct {
-		initialTokensDenomIn  int64
-		initialTokensDenomOut int64
-
-		percentRatio int64
-
-		numShares sdk.Int
-	}
-
-	const (
-		denomOut = "denomOut"
-		denomIn  = "denomIn"
-	)
-
-	now := time.Now().Unix()
-	rng := rand.NewSource(now)
-	t.Logf("Using random source of %d\n", now)
-
-	// generate test case with randomized initial assets and join/exit ratio
-	newCase := func() (tc *testCase) {
-		tc = new(testCase)
-		tc.initialTokensDenomIn = rng.Int63() % 1_000_000
-		tc.initialTokensDenomOut = rng.Int63() % 1_000_000
-
-		// 1%~100% of initial assets
-		tc.percentRatio = rng.Int63()%100 + 1
-
-		return tc
-	}
-
-	swapFeeDec, err := sdk.NewDecFromStr("0")
-	require.NoError(t, err)
-
-	exitFeeDec, err := sdk.NewDecFromStr("0")
-	require.NoError(t, err)
-
-	// create pool with randomized initial token amounts
-	// and randomized ratio of join/exit
-	createPool := func(tc *testCase) (pool *balancer.Pool) {
-		poolAssetOut := balancer.PoolAsset{
-			Token:  sdk.NewInt64Coin(denomOut, tc.initialTokensDenomOut),
-			Weight: sdk.NewInt(5),
-		}
-
-		poolAssetIn := balancer.PoolAsset{
-			Token:  sdk.NewInt64Coin(denomIn, tc.initialTokensDenomIn),
-			Weight: sdk.NewInt(5),
-		}
-
-		pool = createTestPool(t, swapFeeDec, exitFeeDec, poolAssetOut, poolAssetIn).(*balancer.Pool)
-		require.NotNil(t, pool)
-
-		return pool
-	}
-
-	// joins with predetermined ratio
-	joinPool := func(pool types.PoolI, tc *testCase) {
-		tokensIn := sdk.Coins{
-			sdk.NewInt64Coin(denomIn, tc.initialTokensDenomIn*tc.percentRatio/100),
-			sdk.NewInt64Coin(denomOut, tc.initialTokensDenomOut*tc.percentRatio/100),
-		}
-		numShares, err := pool.JoinPool(sdk.Context{}, tokensIn, swapFeeDec)
-		require.NoError(t, err)
-		tc.numShares = numShares
-	}
-
-	// exits for same amount of shares minted
-	exitPool := func(pool types.PoolI, tc *testCase) {
-		_, err := pool.ExitPool(sdk.Context{}, tc.numShares, exitFeeDec)
-		require.NoError(t, err)
-	}
-
-	invariantJoinExitInversePreserve := func(
-		beforeCoins, afterCoins sdk.Coins,
-		beforeShares, afterShares sdk.Int,
-	) {
-		// test token amount has been preserved
-		require.True(t,
-			!beforeCoins.IsAnyGT(afterCoins),
-			"Coins has not been preserved before and after join-exit\nbefore:\t%s\nafter:\t%s",
-			beforeCoins, afterCoins,
-		)
-		// test share amount has been preserved
-		require.True(t,
-			beforeShares.Equal(afterShares),
-			"Shares has not been preserved before and after join-exit\nbefore:\t%s\nafter:\t%s",
-			beforeShares, afterShares,
-		)
-	}
-
-	testPoolInvariants := func() {
-		tc := newCase()
-		pool := createPool(tc)
-		originalCoins, originalShares := pool.GetTotalPoolLiquidity(sdk.Context{}), pool.GetTotalShares()
-		joinPool(pool, tc)
-		exitPool(pool, tc)
-		invariantJoinExitInversePreserve(
-			originalCoins, pool.GetTotalPoolLiquidity(sdk.Context{}),
-			originalShares, pool.GetTotalShares(),
-		)
-	}
-
-	for i := 0; i < 1000; i++ {
-		testPoolInvariants()
 	}
 }
